Move meta and databases to src folder, change require pathing.
