+++ .gitignore
@@ -1,2 +1,2 @@
 node_modules
-meta/token.json
+src/meta/token.json
